Added a debug log parser example
Added an example showing how one might go about extracting data from the
debug log while it's constantly being updated. It also acts as a proxy
server that relays the position information. This is an example only, it
should not be used for production code.

At the time of writing I'm not sure if the 3 values for each position
entry really indicate the X, Y and Z coordinates, will have to get that
verified.